# apps/api/openapi.json
{
  "components": {
    "schemas": {
...
      "Conflict": {
        "additionalProperties": false,
        "description": "One tension between two constraints the writer supplied.\n\n``explanation`` is already rendered: the knowledge base stores a template,\nand the detector substitutes the real values before the conflict is built,\nso nothing downstream needs the knowledge base to display it. ``evidence``\nkeeps those substituted values so a report can be audited without re-running\ndetection.",
        "properties": {
          "evidence": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Evidence",
            "type": "object"
          },
          "explanation": {
            "minLength": 1,
            "title": "Explanation",
            "type": "string"
          },
          "hard_rationale": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hard Rationale"
          },
          "resolutions": {
            "items": {
              "$ref": "#/components/schemas/ResolutionOption"
            },
            "minItems": 2,
            "title": "Resolutions",
            "type": "array"
          },
          "rule_id": {
            "minLength": 1,
            "pattern": "^[a-z][a-z0-9_]*$",
            "title": "Rule Id",
            "type": "string"
          },
          "severity": {
            "$ref": "#/components/schemas/Severity"
          },
          "title": {
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "rule_id",
          "severity",
          "title",
          "explanation",
          "resolutions"
        ],
        "title": "Conflict",
        "type": "object"
      },
...
      "ConstraintBundle": {
        "additionalProperties": false,
        "description": "Everything the writer specified, before any conflict has been detected.\n\nThis is the unit of reproducibility. A conflict report is a pure function\nof a bundle and a knowledge base version, so the pair is enough to explain\nany verdict this system has ever issued.",
        "properties": {
          "audience": {
            "$ref": "#/components/schemas/AudienceSelection"
          },
          "budget_tier_id": {
            "minLength": 1,
            "pattern": "^[a-z][a-z0-9_]*$",
            "title": "Budget Tier Id",
            "type": "string"
          },
          "genre": {
            "$ref": "#/components/schemas/GenreSelection"
          },
          "rating": {
            "$ref": "#/components/schemas/RatingTarget"
          },
          "territories": {
            "$ref": "#/components/schemas/TerritorySet"
          }
        },
        "required": [
          "genre",
          "audience",
          "rating",
          "budget_tier_id",
          "territories"
        ],
        "title": "ConstraintBundle",
        "type": "object"
      },
...
      "ExportBundle": {
        "additionalProperties": false,
        "description": "A project as a self-contained, reproducible document.\n\nCarries the constraint bundle, the resolutions, the envelope and every\nversion that shaped the output. An export that named only the variants\nwould be a set of loglines nobody could defend.",
        "properties": {
          "bundle": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ConstraintBundle"
              },
              {
                "type": "null"
              }
            ]
          },
          "choices": {
            "items": {
              "$ref": "#/components/schemas/ResolutionChoice"
            },
            "title": "Choices",
            "type": "array"
          },
          "conflicts": {
            "items": {
              "$ref": "#/components/schemas/Conflict"
            },
            "title": "Conflicts",
            "type": "array"
          },
          "envelope": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GenerationEnvelope"
              },
              {
                "type": "null"
              }
            ]
          },
          "exported_at": {
            "format": "date-time",
            "title": "Exported At",
            "type": "string"
          },
          "kb_version": {
            "title": "Kb Version",
            "type": "string"
          },
          "markdown": {
            "title": "Markdown",
            "type": "string"
          },
          "project": {
            "$ref": "#/components/schemas/Project"
          },
          "prompt_version": {
            "title": "Prompt Version",
            "type": "string"
          },
          "variants": {
            "items": {
              "$ref": "#/components/schemas/Variant"
            },
            "title": "Variants",
            "type": "array"
          }
        },
        "required": [
          "exported_at",
          "project",
          "kb_version",
          "prompt_version",
          "bundle",
          "conflicts",
          "choices",
          "envelope",
          "variants",
          "markdown"
        ],
        "title": "ExportBundle",
        "type": "object"
      },
...
      "GenerateRequest": {
        "additionalProperties": false,
        "description": "Ask for variants inside a bundle the writer has already resolved.\n\nThe bundle and the choices are sent together rather than the client\nreferencing a stored report. A report is a pure function of a bundle and a\nknowledge base version, so re-deriving it here costs microseconds and\nremoves an entire class of bug where the stored report and the submitted\nbundle have drifted apart.",
        "properties": {
          "bundle": {
            "$ref": "#/components/schemas/ConstraintBundle"
          },
          "choices": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ResolutionChoice"
            },
            "title": "Choices",
            "type": "array"
          },
          "seed": {
            "default": 0,
            "minimum": 0.0,
            "title": "Seed",
            "type": "integer"
          },
          "variant_count": {
            "default": 5,
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Variant Count",
            "type": "integer"
          }
        },
        "required": ["bundle"],
        "title": "GenerateRequest",
        "type": "object"
      },
      "GenerationEnvelope": {
        "additionalProperties": false,
        "description": "The complete, resolved bounds one variant must be generated inside.\n\nLayer 2's output: budget scope bounds, content ceilings reduced to the\nstrictest applicable across every selected territory, the provenance of\neach ceiling, and the guidance the writer's resolutions attached.\n\nIt also names the budget tier and genre it was built from. Those are not\nbounds, but Layer 3 scores archetypes against them, and an envelope that\ncould not say what produced it would force every later stage to carry the\nbundle alongside it and keep the two in step by hand.",
        "properties": {
          "budget_tier_id": {
            "minLength": 1,
            "pattern": "^[a-z][a-z0-9_]*$",
            "title": "Budget Tier Id",
            "type": "string"
          },
          "directives": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/PromptDirective"
            },
            "title": "Directives",
            "type": "array"
          },
          "genre": {
            "$ref": "#/components/schemas/GenreSelection"
          },
          "guidance": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Guidance",
            "type": "array"
          },
          "provenance": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ThresholdSource"
            },
            "title": "Provenance",
            "type": "array"
          },
          "scope": {
            "$ref": "#/components/schemas/ScopeEnvelope"
          },
          "thresholds": {
            "$ref": "#/components/schemas/ContentThresholds"
          }
        },
        "required": ["budget_tier_id", "genre", "scope", "thresholds"],
        "title": "GenerationEnvelope",
        "type": "object"
      },
      "GenerationResponse": {
        "additionalProperties": false,
        "properties": {
          "envelope": {
            "$ref": "#/components/schemas/GenerationEnvelope"
          },
          "failures": {
            "items": {
              "$ref": "#/components/schemas/FailedVariantOut"
            },
            "title": "Failures",
            "type": "array"
          },
          "run": {
            "$ref": "#/components/schemas/GenerationRun"
          },
          "variants": {
            "items": {
              "$ref": "#/components/schemas/Variant"
            },
            "title": "Variants",
            "type": "array"
          }
        },
        "required": ["run", "envelope", "variants", "failures"],
        "title": "GenerationResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Project": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "description",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "Project",
        "type": "object"
      },
      "ProjectCreate": {
        "additionalProperties": false,
        "properties": {
          "description": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "title": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": ["title"],
        "title": "ProjectCreate",
        "type": "object"
      },
      "ProjectList": {
        "additionalProperties": false,
        "properties": {
          "projects": {
            "items": {
              "$ref": "#/components/schemas/Project"
            },
            "title": "Projects",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": ["projects", "total"],
        "title": "ProjectList",
        "type": "object"
      },
      "ProjectUpdate": {
        "additionalProperties": false,
        "description": "A partial update. Every field optional; omitted means unchanged.\n\nDistinct from :class:`ProjectCreate` so that omitting a title on a PATCH is\n\"leave it alone\" rather than \"clear it\", which a shared model could not\nexpress.",
        "properties": {
          "description": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "title": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "title": "ProjectUpdate",
        "type": "object"
      },
...
      "ResolutionChoice": {
        "additionalProperties": false,
        "description": "The writer's answer to one conflict.",
        "properties": {
          "resolution_id": {
            "minLength": 1,
            "pattern": "^[a-z][a-z0-9_]*$",
            "title": "Resolution Id",
            "type": "string"
          },
          "rule_id": {
            "minLength": 1,
            "pattern": "^[a-z][a-z0-9_]*$",
            "title": "Rule Id",
            "type": "string"
          }
        },
        "required": ["rule_id", "resolution_id"],
        "title": "ResolutionChoice",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": ["loc", "msg", "type"],
        "title": "ValidationError",
        "type": "object"
      },
...
      "VariantList": {
        "additionalProperties": false,
        "properties": {
          "total": {
            "title": "Total",
            "type": "integer"
          },
          "variants": {
            "items": {
              "$ref": "#/components/schemas/Variant"
            },
            "title": "Variants",
            "type": "array"
          }
        },
        "required": ["variants", "total"],
        "title": "VariantList",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "description": "Supabase access token",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Constraint-Aware Script Ideation Engine — deterministic constraint reasoning and archetype-forced plot variant generation.",
    "title": "ScriptGenie API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/projects": {
      "get": {
        "operationId": "list_projects_v1_projects_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List your projects",
        "tags": ["projects"]
      },
      "post": {
        "operationId": "create_project_v1_projects_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create a project",
        "tags": ["projects"]
      }
    },
    "/v1/projects/{project_id}": {
      "delete": {
        "operationId": "delete_project_v1_projects__project_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete a project",
        "tags": ["projects"]
      },
      "get": {
        "operationId": "read_project_v1_projects__project_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Read one project",
        "tags": ["projects"]
      },
      "patch": {
        "operationId": "update_project_v1_projects__project_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update a project",
        "tags": ["projects"]
      }
    },
    "/v1/projects/{project_id}/export": {
      "get": {
        "description": "The project as JSON and Markdown, with everything behind it.\n\nBoth formats in one response rather than two endpoints, because they are\ntwo renderings of one document and returning them together is what\nguarantees they cannot disagree.",
        "operationId": "export_project_v1_projects__project_id__export_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportBundle"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Export a project with its full provenance",
        "tags": ["projects"]
      }
    },
    "/v1/projects/{project_id}/generate": {
      "post": {
        "description": "Run the pipeline. Blocked while any HARD conflict is unresolved.\n\nThe order matters. The bundle is detected and resolved first —\ndeterministic, free, and the step that raises 409 with the blocking\nconflicts attached — so a bundle that cannot legally be generated never\nreaches the model.",
        "operationId": "generate_v1_projects__project_id__generate_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Generate plot variants for a resolved bundle",
        "tags": ["projects"]
      }
    },
    "/v1/projects/{project_id}/variants": {
      "get": {
        "operationId": "list_variants_v1_projects__project_id__variants_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VariantList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List a project's generated variants",
        "tags": ["projects"]
      }
    },
...
  }
}